the fix for known/unknown child UTIs to be handled by QLStephen
So cheaply pretending to be a QLGenerator from Apple seems to be the way
to make Quick Look use QLStephen also for UTIs with direct, indirect,
explicit and implicit (for dynamic UTIs) conformance to UTI registered
for it.
If this is the fix, I can only remember that most used language in
programming is profanity.